Tweak test organisation
