+++ tests/test_proposal/test_flowproposal/test_flowproposal_properties.py
